feat(KRV-02): create authentication with symfony security
Create authentication inside of system Korv to improve access only users
autorizate to use that application. We use Symfony Security to provide
permissions in specific routes and Lexik_JWT to generate a hash valid
the current user to access that context.

Obs: We need to construct "forgot my password?" and prolong how much
time that hash is valid.

Resolves: KRV-03, KRV-04, KRV-05, KRV-06