feat(KRV-27): update login route and create logout route
Update login route because front-end needs a specific payload when user
is authenticated. That is email, roles and token (continued created with
JWT Levik following Sympony documentation) to valid that user in front.

Create logout route because I have a lot of problems to understood how
can I use login/logout in Sympony, but, now is clearly for me XD.